Move people.json to a more sensible location Fixes #140
